requirements.txt
# Core dependencies for Vietnam Stock AI Backend
# Compatible with Python 3.10.12 (Anaconda stock environment)

# Data collection
vnstock==3.3.0
kafka-python==2.0.2

# Database
pymongo==4.6.1

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.12.5
email-validator>=2.0.0

# Workflow orchestration
apache-airflow==2.7.3

# Data processing
pandas==2.1.4
numpy==1.26.4

# AI/ML
scikit-learn==1.3.2
ta==0.11.0

# LLM integration
openai==1.3.0
anthropic==0.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
hypothesis==6.92.0
mongomock==4.1.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
